move MintNFT to component folder from utils
